errors: Add a report for line-errors.
Sometimes it is more useful  or realistic to say that an entire line caused an error than to say a specific token caused an error. `LineReport` specifically targets this usecase.

This commit also introduces a `Report` type that all future errors in the project should implement.

+++ ti-basic-optimizer/src/parse/expression.rs
@@ -48,7 +48,7 @@ impl<'a> Builder<'a> {
         self.finalize()
     }
 
-    fn error(&self, code: usize) -> TokenReport {
+    fn error(&self, code: u16) -> TokenReport {
         TokenReport::new(
             self.tokens.current_position(),
             "Expression parsing error",
